+++ base/views/learning_unit.py
@@ -23,22 +23,17 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-import itertools
 from copy import deepcopy
 from decimal import Decimal
 
 from django.conf import settings
 from django.contrib.auth.decorators import login_required, permission_required
-from django.db.models import Sum
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, render
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _, get_language
 from django.views.decorators.http import require_http_methods
 
-from attribution.business import attribution_charge_new
-from attribution.models.attribution_charge_new import AttributionChargeNew
-from attribution.models.enums.function import Functions
 from base import models as mdl
 from base.business.learning_unit import get_cms_label_data, \
     get_same_container_year_components, CMS_LABEL_SPECIFICATIONS, get_achievements_group_by_language, \
@@ -54,7 +49,6 @@ from base.forms.learning_unit_specifications import LearningUnitSpecificationsFo
 from base.models import education_group_year, proposal_learning_unit
 from base.models import learning_component_year as mdl_learning_component_year
 from base.models.entity_version import EntityVersion
-from base.models.enums import learning_unit_year_subtypes
 from base.models.enums.attribution_procedure import ATTRIBUTION_PROCEDURES
 from base.models.enums.entity_container_year_link_type import EntityContainerYearLinkTypes
 from base.models.enums.learning_component_year_type import LEARNING_COMPONENT_YEAR_TYPES
@@ -62,7 +56,7 @@ from base.models.enums.learning_unit_year_periodicity import PERIODICITY_TYPES
 from base.models.enums.vacant_declaration_type import DECLARATION_TYPE
 from base.models.learning_unit_year import LearningUnitYear
 from base.models.person import Person
-from base.views.common import display_warning_messages, display_success_messages
+from base.views.common import display_success_messages
 from base.views.learning_units.common import get_common_context_learning_unit_year, get_text_label_translated
 from cms.models import text_label
 from cms.models.translated_text_label import TranslatedTextLabel
@@ -110,23 +104,6 @@ def learning_unit_components(request, learning_unit_year_id):
     context['tab_active'] = 'components'
     context['can_manage_volume'] = business_perms.is_eligible_for_modification(context["learning_unit_year"], person)
     return render(request, "learning_unit/components.html", context)
-
-
-@login_required
-@permission_required('base.can_access_learningunit', raise_exception=True)
-def learning_unit_attributions(request, learning_unit_year_id):
-    context = get_common_context_learning_unit_year(learning_unit_year_id, request.user.person)
-
-    context['attributions'] = attribution_charge_new.find_attributions_with_charges(learning_unit_year_id)
-    context["can_manage_charge_repartition"] = business_perms.is_eligible_to_manage_charge_repartition(
-        context["learning_unit_year"], request.user.person
-    )
-    context["can_manage_attribution"] = business_perms.is_eligible_to_manage_attributions(
-        context["learning_unit_year"], request.user.person
-    )
-    warning_msgs = get_charge_repartition_warning_messages(context["learning_unit_year"].learning_container_year)
-    display_warning_messages(request, warning_msgs)
-    return render(request, "learning_unit/attributions.html", context)
 
 
 @login_required
@@ -474,44 +451,6 @@ def _get_learning_unit_year(academic_yr, learning_unit_yr):
     if learning_unit_years.exists():
         return learning_unit_years.first()
     return None
-
-
-def get_charge_repartition_warning_messages(learning_container_year):
-    total_charges_by_attribution_and_learning_subtype = AttributionChargeNew.objects \
-        .filter(attribution__learning_container_year=learning_container_year) \
-        .order_by("attribution__tutor", "attribution__function", "attribution__start_year") \
-        .values("attribution__tutor", "attribution__tutor__person__first_name",
-                "attribution__tutor__person__middle_name", "attribution__tutor__person__last_name",
-                "attribution__function", "attribution__start_year",
-                "learning_component_year__learning_unit_year__subtype") \
-        .annotate(total_volume=Sum("allocation_charge"))
-
-    charges_by_attribution = itertools.groupby(total_charges_by_attribution_and_learning_subtype,
-                                               lambda rec: "{}_{}_{}".format(rec["attribution__tutor"],
-                                                                             rec["attribution__start_year"],
-                                                                             rec["attribution__function"]))
-    msgs = []
-    for attribution_key, charges in charges_by_attribution:
-        charges = list(charges)
-        subtype_key = "learning_component_year__learning_unit_year__subtype"
-        full_total_charges = next(
-            (charge["total_volume"] for charge in charges if charge[subtype_key] == learning_unit_year_subtypes.FULL),
-            0)
-        partim_total_charges = next(
-            (charge["total_volume"] for charge in charges if charge[subtype_key] == learning_unit_year_subtypes.PARTIM),
-            0)
-        partim_total_charges = partim_total_charges or 0
-        full_total_charges = full_total_charges or 0
-        if partim_total_charges > full_total_charges:
-            tutor_name = Person.get_str(charges[0]["attribution__tutor__person__first_name"],
-                                        charges[0]["attribution__tutor__person__middle_name"],
-                                        charges[0]["attribution__tutor__person__last_name"])
-            tutor_name_with_function = "{} ({})".format(tutor_name,
-                                                        getattr(Functions, charges[0]["attribution__function"]).value)
-            msg = _("The sum of volumes for the partims for professor %(tutor)s is superior to the "
-                    "volume of parent learning unit for this professor") % {"tutor": tutor_name_with_function}
-            msgs.append(msg)
-    return msgs
 
 
 def get_specifications_context(learning_unit_year, request):
